Add PwnContext and move Config to it's own package
